pyproject.toml
[project]
name = "python-template-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "flake8>=7.3.0",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.16.2",
    "uvicorn>=0.35.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --tb=short"
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s"
log_file = "test_logs/pytest.log"
log_file_level = "INFO"

[build-system]
requires = ["setuptools>=65.5.0"]
build-backend = "setuptools.build_meta"
